+++ bin/main.d.ts
@@ -1,4 +1,5 @@
 /**
  *  Run end to end Tableland
  **/
+export declare const shutdown: () => Promise<never>;
 export declare const main: () => Promise<void>;
